Split the code into three separate files
Split the code so that it becomes easier to read and to make
it easier to update it in the future. Update package names
and edit files to follow Java standards.
