+++ examples/java/basic_example/.gitignore
@@ -1,0 +1,3 @@
+dependency-reduced-pom.xml
+/target/*
+!/target/javapluginlogger.properties
