+++ examples/java/basic_example/SSEJavaPlugin.qvf
